chore: extract dates badge
